feat(css): Spring is a real AnimationInterpolationFunction variant
Retires the `Interp` wrapper's reason to exist. A spring is now expressible as a
timing function like any other, rather than a Rust-only concept bolted beside
the enum:

    pub enum AnimationInterpolationFunction {
        Ease, Linear, EaseIn, EaseOut, EaseInOut,
        CubicBezier(SvgCubicCurve),
        Spring(SpringCurve),          // new
    }

`SpringCurve` lives in azul-css next to the enum that carries it, and azul-core
re-exports it as `Spring`. One definition, not two structurally-identical
`#[repr(C)]` types that would become a silent ABI trap the day either gained a
field. The spring maths moved with the type; the constants are byte-identical to
core's previous ones (GENTLE 120/20, SNAPPY 260/20 — NOT the values I first
wrote in css, which would have quietly changed how every default transition
feels).

A spring is unlike every other variant in one way, and both new accessors exist
to stop that being a trap:

  * `is_spring()` — a spring has NO duration and cannot be evaluated at a
    normalised `t`. Anything that owns a timeline must branch on this.
  * `get_curve()` / `ease()` return an ease-in-out stand-in rather than
    panicking, and they agree with each other, so a caller that can only think
    in curves (CSS serialisation, the SVG preview) degrades to plausible motion.
    `AnimChannel` never takes that path — it dispatches on `Interp` and
    integrates `Spring::step` from the live (value, velocity).

NO api.json change and no codegen run, contrary to what the plan assumed: this
enum is not in api.json at all (0 occurrences; its payload types SvgCubicCurve
and SvgPoint are there for unrelated reasons). It was never part of the exposed
C surface, so nothing regenerates and no binding breaks. Exposing timing
functions to the bindings is a separate decision and is not needed for the
engine-driven model, where the app describes the DOM and does not configure
animations.

The module docs that justified wrapping-instead-of-extending are rewritten
rather than left behind — they now describe what the code does.

Verified: azul-css 2845/0, azul-core animation 24/24 (including a new test
pinning the variant through the enum's own accessors), azul-layout --all-targets
and azul-dll (link-static,pdf,debug-server) check clean.
